requirements.txt
python-telegram-bot==20.7
openai==1.52.0
nest_asyncio
google-auth
google-auth-oauthlib
google-api-python-client
